refactor(frontend): extract loading logic, change rendering
